refactor: OrderTable을 Order 내부로 변경
