+++ src/main/java/kitchenpos/tablegroup/GroupedOrderTables.java
@@ -3,7 +3,7 @@ package kitchenpos.tablegroup;
 import java.util.List;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
-import kitchenpos.ordertable.OrderTable;
+import kitchenpos.order.OrderTable;
 
 public class GroupedOrderTables {
