+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -5,9 +5,9 @@ import static java.util.stream.Collectors.toList;
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.order.Order;
+import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.ordertable.OrderTable;
-import kitchenpos.ordertable.persistence.OrderTableRepository;
+import kitchenpos.order.persistence.OrderTableRepository;
 import kitchenpos.tablegroup.TableGroup;
 import kitchenpos.tablegroup.persistence.TableGroupRepository;
 import kitchenpos.tablegroup.request.OrderTableDto;
