+++ src/main/java/kitchenpos/ui/TableRestController.java
@@ -2,9 +2,9 @@ package kitchenpos.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.ordertable.OrderTable;
-import kitchenpos.ordertable.application.OrderTableService;
-import kitchenpos.ordertable.request.OrderTableCreateRequest;
+import kitchenpos.order.OrderTable;
+import kitchenpos.order.application.OrderTableService;
+import kitchenpos.order.request.OrderTableCreateRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
