+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -19,11 +19,11 @@ import kitchenpos.menu.persistence.MenuRepository;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderStatus;
+import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
+import kitchenpos.order.persistence.OrderTableRepository;
 import kitchenpos.order.request.OrderCreateRequest;
 import kitchenpos.order.request.OrderLineItemCreateRequest;
-import kitchenpos.ordertable.OrderTable;
-import kitchenpos.ordertable.persistence.OrderTableRepository;
 import kitchenpos.product.Product;
 import kitchenpos.product.persistence.ProductRepository;
 import kitchenpos.support.ServiceTest;
