+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -16,9 +16,9 @@ import kitchenpos.menu.persistence.MenuRepository;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderStatus;
+import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.ordertable.OrderTable;
-import kitchenpos.ordertable.persistence.OrderTableRepository;
+import kitchenpos.order.persistence.OrderTableRepository;
 import kitchenpos.product.Product;
 import kitchenpos.product.persistence.ProductRepository;
 import kitchenpos.support.ServiceTest;
